Settings update across instances, more small fixes

+++ Source/Pd/PdGui.h
@@ -119,7 +119,7 @@ class Gui : public Object
     
     // pd expresses some object's width in characters
     // This variable sets the conversion ratio from chars to width
-    static inline constexpr int nbxCharWidth = 12;
+    static inline constexpr float nbxCharWidth = 12.0f;
 };
 
 }  // namespace pd
